fix some bugs
1 fix the cpu path bug of select
2 fix the bug of reduce
3 optimzation of the reduce

case:
Onnx_Where_test_basic_dfp16
Onnx_Where_test_broadcast_diverse_dims_dfp16
Onnx_Where_test_broadcast_same_dims_dfp16
TF_Where_test_with_var_dfp16
TF_ReduceSum_test_multiple_axis
Pytorch_Sum_test_basic
Pytorch_Sum_test_dimlist

+++ ovxlib/src/kernel/cpu/select_cpu.c
@@ -137,7 +137,7 @@ DEF_KERNEL_EXECUTOR(_compute)
         in2_offset = _expand_offset( i, in_attr[2]->shape->data, in_attr[2]->shape->size,
                 in_stride_size[2], out_attr[0]->shape->data );
 
-        f32_out_buffer[0][i] = ((int32_t)(f32_in_buffer[0][in0_offset])) ?
+        f32_out_buffer[0][i] = (f32_in_buffer[0][in0_offset]) ?
                                f32_in_buffer[1][in1_offset] : f32_in_buffer[2][in2_offset];
     }
